+++ workspaces/bazaar/plugins/bazaar/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @backstage-community/plugin-bazaar
 
+## 0.6.0
+
+### Minor Changes
+
+- 9e1fa3b: Backstage version bump to v1.34.1
+
 ## 0.5.2
 
 ### Patch Changes